Introduce scalaShadowing package
The `scalaShadowing` package is used to safely modify classes and
objects in scala so that they can be used from dotty. They will
be visible as members of the `scala` package, replacing any objects
or classes with the same name. But their binary artifacts are
in `scalaShadowing` so they don't clash with the same-named `scala`
members at runtime.