Plug another space leak.
`initInfo` was retained in Symbols. When called from `Namer`, `initInfo` referred to
a completer, which referred to a context.

With this space leak plugged, we can now compile 1000 times core/Comments.scala (460lines)
with -Xmx400M.

There still seems to be a space leak on the order of 200KB per run, though. But that seems
to have to do with symbols, not contexts.