Make tests depend on environment variables for classpath jars
sbt adds the correct jars to classpath and the tests depend on
`packageAll` which creates these. When using something else however,
these together with `sbt-interfaces` do not get propagated from the
build.

To remedy this and make the testing a bit more flexible, we now
take these from `sys.props` instead, see `tests/dotty/Jars.scala`.

If the props aren't defined we fall back to the ones default to sbt.