Modules w. serializable type alias "companions" are not serializable
The behaviour changed in scala/scala#5550, this commit adapts to the change so
that we'll be binary compatible after boostrapping.

MiMa alerted us to a change in the parentage of two objects in the
forkjoin package object.

In Scala 2.12.0/1, they implemented `scala.Serializable`. Recently,
this (synthetically added) parent was absent. This appears to be
due to a bug fix in `companionSymbolOf`, which no longer treats
objects and same-named type aliases to be companions.

This commit manually adds the formerly-synthetic parents to these
objects, and documents the change in compiler behaviour with a test.

Fixes scala/scala-dev#290