Fix #1797: Allow case class params with names _1, _2, ...
This was not possible before because it clashed with the automatically
generated name of the accessor. We now allow it, by simply taking the
parameter(accessor) itself as the case class accessor if it already has
that name. But you still cannot write

    case class C(_2: Int, _1: String)

nor should you be able to do this.