Adopt scala's scheme for root import hiding
scalac hides a root import from Predef if there is an eplicit Predef import.
We now do the same (previously we did this only if the overriding import undefined
something, using a `x => _` syntax). To avoid cycles and races one had to be very careful
not to force import symbols too early, so we now compare the name before the symbol proper.

All this is likely temporary - the comment of ImportInfo#unimported points to a different,
more systematic solution.