Fix #1905: Duplicate bridge
The problem with #1905 was that we checked "new" overriding relationships
are phase erasure + 1. This is wrong when we have ErasedValueTypes because
these do not compare correctly ith their underlying type. So we get spurious
mismatches which force in turn spurious bridge generation. The fix is to
compute newly overridden symbols at phase elimErasedValueTypes + 1, i.e. one
phase later.