Remove the error recovery mechanism in RedBlackTree.TreeIterator.
After fixing the stack size computation in
efc8550832aee955a481b5568bee386b26062460, this commit removes the
error recovery mechanism which hid the internal bug. As the
comment in `pushNext` says, "Obviously this should never happen".

This commit is potentially risky, if we still do wrongly compute
the maximum stack size, but further problems are not caught by
the test suite. But at least we would now know when we fail to
compute the stack size correctly, instead of letting it go
unnoticed for years.

This change also ensures that Scala/JVM does not silently succeed
where Scala.js would fail.