Weaken assumption in mergeCompanionDefs
Previously `mergeCompanionDefs` assumes that if the attachment of class `Foo` is as follows:

    x :: y :: tdef @ TypeDef(_, templ)

Then the `tdef` must be `Foo$`. When there are multiple pre-typer transforms,
this is not necessarily true. For example, an annotation macro expansion may
expand a non-case class `Foo` to `class Foo; object FooA`.

We need to check the name of `tdef` to be equal to `Foo$`.