during bootstrap, don't build scala-parser-combinators or -swing
I see no further reason to include them in the bootstrap process.
Adriaan agrees (in-person discussion).

note that the community build makes sure the modules still compile as
we make changes to Scala.

as for actually publishing scala-parser-combinators and scala-swing
when we make binary incompatible changes to Scala, I think it's fine
to consider that properly the job of the module maintainers, rather
than properly the job of this script.  (if we want more automation
on that, we could make some elsewhere, keeping that concern separate
from actual bootstrapping concerns.)

scala-xml, scala-partest, and scalacheck (commented out at the moment)
remain in this script since they are all actually still part of the
bootstrapping picture.

fixes https://github.com/scala/scala-dev/issues/302 , so that
integrate-bootstrap will pass again (currently it fails, since
scala-parser-combinators merged Scala.js support)