make Predef.*CharSequence implicits public again
fixing a recent regression in 2.13

when I was removing deprecated stuff in
https://github.com/scala/scala/pull/5683/
I was going too fast and misunderstood these deprecations.
only the names of the class parameters were supposed to become
private, not the entire class.

note that now that the parameters aren't even vals anymore, they don't
need to have funny `__` names anymore, either.

this was caught by the Scala community build.