Better fix for ParamForwarder under non-separate compilation
We cannot rely on ParamForwarder having been run on the superclass even
if we're in the correct phase since ParamForwarder is not a mini-phase.
Instead, we should accept a candidate even if it's not a method.

This fixes `tests/run/paramForwarding_together_b.scala`, this also required
updating `tests/run/paramForwarding.scala` since the field in `Y` is not
actually required.