Update tests after tuple changes
- dotc/scala-collections.blacklist:

Blacklist Unit and TupleN from scala-collections test

- tests/neg/i1653.scala:

Having an additional superclass on Unit seams to affect the error here (not sure why).

- tests/pos/t0301.scala:

Since `_2` is now added thought implicit conversion, the pattern used in this case is no longer stable.

- tests/pos/tcpoly_bounds1.scala:

This test uses the fact that `(A, B) <: Product2[A, B]`, which is no longer the case.

- tests/repl/patdef.check:

Having the `_i` via implicit conversions remove the @unchecked here

- tests/run/tuple-match.scala:

Replace Tuple1 in pattern matching with explicit TupleCons pattern. TupleN can still be used in apply position (defined as defs in DottyPredef), but not in unapply position.

- tests/run/unapply.scala:

Also uses the fact that `(A, B) <: Product2[A, B]`.

- tests/run/withIndex.scala:

Might be a limitation in typer?:

14 |        case _: Array[Tuple2[_, _]] => true
   |                      ^^^^^^^^^^^^
   |unreducible application of higher-kinded type [A, B] => dotty.TupleCons[A, dotty.TupleCons[B, Unit]] to wildcard arguments