Erasure/Scala2Unpickler hacks to restore binary compatibility
To stay binary compatibility with scala 2.11 binaries, we hijack the
unpickling of types to fake an HList structure. Later on in erasure, HList
types are eraised back to scala.TupleN (for N <= 22). In addition because the
actual scala.TupleN classes are neither `<: Tuple` now `<: TupleCons`, these
types needs to be systematically araised to Object.

In addition to all these carefully compensating hacks, this also imposes a new
contain: the dotty-library, which defines several `Tuple/TupleCons` methods,
can should now *always* be compiled by dotty itself. Indeed, when compiled
with scalac, these types will stay as they are instead of being eraised to
Object.