Make forwardTypeParams configurable
A new setting, -Ysuppress-param-forwarding is used
to make forwarding type parameters confihurable.

Forwarding type parameters from base class to subclass instance
is done in `normalizeToClassRefs` for optimization, but I just found
out this optimization is not always type-preserving and in fact can
lead to type errors.

SortedMap.scala in the collection strawman is an example. With parameter
forwarding enabled, we get:

    Error: immutable/SortedMap.scala --------------------------------------------
    |    with MapValuePolyTransforms[K, V, C] {
    |                                      ^
    |Type argument C does not conform to upper bound [=X, +Y] =>
    |  strawman.collection.immutable.Map[=X, +Y] &
    |    strawman.collection.immutable.MapValuePolyTransforms[=X', +Y', LazyRef(C)]
    |
    |where:    +Y  is a type variable
    |          +Y' is a type variable
    |          =X  is a type variable
    |          =X' is a type variable

If we turn off forwarding with the option, the strawman compiles OK.

Unfortunately turning the option off altogether does not work (yet).
It leads to increased compile times, and more tests fail with deep subtype
recursions. Dotty bootstrap seems to not terminate at all.

Related: Refine notion of self in RefChecks. This is
Needed to make pos/i1401.scala compile when forwardTypeParams is off.