Make the sbt scripted tests use the bootstrapped dotty
Also move them to their own project `dotty-sbt-scripted-tests`, because
they need to be in a Scala 2.10 project to work correctly, since sbt
uses 2.10.

The scripted tests can now be run with:
> dotty-sbt-scripted-tests/scripted

A few of them are broken right now, I'll fix them later (we don't run
them on the CI yet).