New bridge implementation
This one is based on the way things are done in scalac.

As yet missing: Systematic treatment of errors.

Test i1240a was moved to pending. It caused a merge error before
but this was likely accidental. We have to check whether systematic
bridge error checking will discover a problem. For the moment
the test behaves exactly as for scalac: it compiles and the result
is the same.