Merge prepareIDE with configureIDE
In the next commit we'll extend configureIDE to work even when it needs
to update some settings to set the scalaVersion to Dotty. Since this
update takes some time and is discarded at the end of the task, it makes
sense to run the configuration and compilation steps in the same task to
avoid having to do the settings update twice.