Reject trailing `_` syntax under -strict mode.
The reason to do this only under -strict mode for now is so that
we can still compile the same code under both Scala 2 and Dotty.
If we rejected trailing `_` outright, Scala 2 code would need to
get either explicitly eta-expanded or get an expected type to still work.
But in Dotty, we can simply drop `_` for functions with arity >= 1.
So to keep code dual compilable, we'd need to add boilerplate
which under Dotty would be no longer needed.