Fix #2812: update author-image width
Fixing author-image to be bound to 100px max. Fixes: #2812

With:
```
width:100px
height:auto
```

On firefox, the image is scaled to be 100x100. However on chrome, it ends up being 200x100.

Since we wan't to fix the width to 100px (from that original css), I've updated the max-width and max-height to be 100, so the image is always squared.