Fix positions in assignments
TreeCopier#finalize will set the position of the new tree to the old
tree, which is not what we want here since the desugared tree position
should contain more than the position of the parsed assignment lhs.
Moreover, the desugared tree will never be identical to the parsed lhs
so there is no reason to use a TreeCopier at all.