Avoid unnecessary narrowLiftedOwner in markFree
This prevented many local functions from being lifted out
to static scope, and thereby caused memory leaks. The memory leak found
in the IDE was caused an anonymous function in NamerContextOps (a base
trait of Context) that was kept as a non-private private method of
NamerContextOps, taking a `this` referring to a context.