Fix #2300: Try to heal subtype failures by instantiating type variables
The idea is that if an adapt fails because an actual type A is not a
subtype of an expected type B, we try to instantiate all type variables
in A and, if that changes something, try again.

This mitigates the discrepancy that dotc is more lazy in instantiating type
variables than scalac.

Fixes #2300.