Fix crash when comparing HKTypeLambda and RefinedType
Calling HKTypeLambda#member will crash when trying to form a NamedType,
this happened in `compareRefinedSlow` so we now skip it. If this
HKTypeLambda is an eta-expansion it will be handled by `fourthTry`.

Somehow, this only started happening after the last commit, when
running:
> vulpix tests/neg
But it's not deterministic, multiple runs may be necessary.
The crash happened while compiling i1650.scala, but compiling this file
by itself didn't trigger the failure.