Check that class type parameters are only referenced via their this-types
Previously, class type parameters were public, and thus could be referenced
with any prefix. Now they are private local. The check ensures that transforms
respect that invariant and that we know as soon as possible if they don't.

The flag is normally disabled, because there are still some legitimate cases
that trigger it. Need to investigate these further.