Refactor test
Legacy-tests seems to always compile files together, but cooking
only works for Java files read by ClassfileParser. I don't think
it makes sense to change that - if we compile Java files from sources
we should demand that raw types are re-written as wildcard types.

So to fix this we move the offending files to pos-special and compile
them only in CompilationTests.