Fix #2732: Allow wildcards in SAM types
The `isInstantiable` test as written is incorrect:
Given tp = `java.util.function.Function[String, _ <: String]`, we will have:
    tref = skolem TermRef with underlying info `tp`
    selfType = `java.util.function.Function[String, tref#R]` (because of the `asSeenFrom`)
`tref <:< selfType` is false, because `tref.underlying` is
just `tp` and its return type parameter is a bounded wildcard, not a reference with
`tref` as a prefix.
In any case, I cannot think of a case where this function should return
false, and no test failed when I removed it. So I'm removing it until
someone can come up with a counter-example.