Do relaxedTyping everywhere
If all phases run at group end, any phase that adds or removes method parameters
makes all phases run before it in the same group have the wrong number of parameters
when type assigning an Apply or TypeApply. Previously this was allowed only if
relaxedTyping was true. Now it is allowed everywhere. We should still be able to
catch any errors in the transformer logic using Ycheck.

With this change, LinkScala2Impls can now be run at group end.