Deprecate ClassfileAnnotation, introduce ConstantAnnotation
Annotations that need to be written to the classfile in the
Java-compatible format need to be defined in Java.

Scala annotations extending ConstantAnnotation can only accept
compile-time constants as argument. The implementation represents
constant annotations using the `assocs` field in the AnnotationInfo,
just like for Java-defined annotations. The implementation for getting
the constants and building the `assocs` is the same as for Java
annotations, and the same restrictions apply: arguments need to be
passed as named arguments (except if there's a single argument defining
the `value` parameter). This is a bit of an ad-hoc restriction for
Scala-defined constant annotations, but changing it to accept positional
args went over my time budget (there's no easy way to reuse the code
that deals with named arguments in general).

Internally, we still add the `StaticAnnotation` parent to Java-defined
annotations (in the Classfile parser, Java source parser, runtime
reflection) to make sure that Java annotations are not only emitted to
the classfile in the Java format, but also pickled.