Fix #3352: defaults parameters require named type in compatibility check
In implicit search, if the expected type is applying type and the candidate
is a TermRef to a method with default parameters, we need the TermRef
to the method in order for `TestApplication` to insert default parameters.
Otherwise, by widening the TermRef to MethodType, the information about default
parameters is lost,  which may incorrectly disqualify a candidate.