Remove compiler dependency on `@BeanInfo`
The `BeanInfo` generation was deprecated in 2.12 and will be removed in 2.13.
First, we cut ties from the compiler in M3, so that we can remove the
classes `scala.beans.{BeanInfo, ScalaBeanInfo}` entirely in M4, along with
`scala.beans.{BeanInfoSkip, BeanDescription, BeanDisplayName}`, which
were part of the deprecated feature. (Note: for good measure, we should
deprecate all of them in 2.12.5, instead of just the entry point).

This cuts tie 2 of 3 to the Java profiles beyond compact1.