Avoid loading compilation units twice
For some class name foo.Foo we need to try to load the tasty from the its class.
If the class is not definded, we try load it from the obeject.

Previously, if both the class and the object extists we were loading the compilation unit twice.
Which would duplicate symbols and then fail with duplicated symbol definitions.