Fix problem with partly uninstantiated dependent functions
If the result type of a dependent function contains a type variable whose bounds
refer to the function's parameters, that type variable has to be instantiated before
the method type is formed. Otherwise, if the type variable is instantiated later
the method result type will refer to the original parameter symbols instead of
the method types ParamRefs.