Fix Erasure's transformDenot for SingleDenotations
They wold transform the pre-existing type of the denotation instead of the
symbol's type. This is clearly the wrong behavior for erasure. It was not
detected before because we changed all types from name-based to symbolic
during erasure, which meant we computed the denotation after erasure by a
loadDenot. So we never exercised the `d.current` code on a derived denotation.