Always pickle the position of DefTrees
A symbol is created before the corresponding tree is unpickled,
so we cannot rely on initialPos to set the symbol position.
Instead, we always pickle the position of definitions.

In practice, this means that we now always pickle the position of all
`TypeDef` and `Bind` (the other `DefTree`s are all `WithLazyField` and
thus already had their positions pickled).